# dlp-server/openapi.yaml
openapi: "3.1.0"
info:
  title: GhostLayer DLP API
  version: "3.0.0"
  description: |
    **GhostLayer** is an enterprise Data Loss Prevention (DLP) platform.

    This specification documents the REST APIs used to provision tenants,
    generate agent API keys, and poll or stream security events.  All
    endpoints require super-admin authentication via the
    `x-super-admin-key` request header (or a valid `super_admin_auth`
    session cookie for browser-based flows).

    ## Authentication
    Supply your super-admin key in every request:
    ```
    x-super-admin-key: <YOUR_SUPER_ADMIN_KEY>
    ```

    ## Infrastructure-as-Code (IaC)
    Use the `/api/provision` endpoint together with `/api/tenants` to
    fully automate GhostLayer deployments from Terraform, Ansible, or
    any CI/CD pipeline.

  contact:
    name: GhostLayer Security Team
    url: https://ghostlayer.io
  license:
    name: Proprietary
servers:
  - url: "https://{serverUrl}"
    description: Configurable server endpoint
    variables:
      serverUrl:
        default: localhost:3000
        description: API host and port (without protocol)
  - url: http://localhost:3000
    description: Local development

# ── Security ───────────────────────────────────────────────────────────────────
security:
  - SuperAdminKey: []
components:
  securitySchemes:
    SuperAdminKey:
      type: apiKey
      in: header
      name: x-super-admin-key
      description: Super-admin API key set via the `SUPER_ADMIN_KEY` environment variable.

  # ── Reusable schemas ─────────────────────────────────────────────────────────
  schemas:
...
    Agent:
      type: object
      description: A GhostLayer agent instance deployed inside a tenant network.
      required: [_id, tenantId, name, agentKey, environment, createdAt]
      properties:
        _id:
          type: string
        tenantId:
          type: string
          description: ObjectId of the parent tenant.
        name:
          type: string
          examples: ["london-office-agent"]
        agentKey:
          type: string
          description: Unique key used by the agent to authenticate heartbeats.
        environment:
          type: string
          enum: [production, staging, development]
          default: production
        lastPing:
          type: string
          format: date-time
          description: UTC timestamp of the most-recent heartbeat.
        createdAt:
          type: string
          format: date-time
    DeploymentConfig:
      type: object
      description: Ready-to-use deployment artefacts for the new agent.
      properties:
        envConfig:
          type: string
          description: |
            Newline-delimited environment variable assignments suitable for a
            `.env` file or a systemd `EnvironmentFile=` directive.
          examples:
            - |
              DLP_SERVER_URL=https://your-dlp-server.example.com
              DLP_TENANT_API_KEY=<uuid>
              DLP_AGENT_KEY=agent-abc123
              DLP_ENVIRONMENT=production
              DLP_PING_INTERVAL=30000
        dockerCommand:
          type: string
          description: |
            Fully-parameterised `docker run` command that starts the agent
            container with the correct environment variables.

    # ── Events ───────────────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          examples: ["name and contactEmail are required"]
...
  responses:
    Unauthorized:
      description: Missing or invalid super-admin credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /api/provision:
    post:
      operationId: provisionAgent
      summary: Provision a new agent and generate its API key
      description: |
        Creates a new GhostLayer agent record for an existing tenant and
        returns a unique `agentKey` together with ready-to-use deployment
        artefacts (`.env` snippet and `docker run` command).

        This is the primary **API key generation** endpoint for IaC workflows.
        Each call produces one agent with a unique, non-reusable key.
      tags: [Agent Provisioning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, name]
              properties:
                tenantId:
                  type: string
                  description: ObjectId of the tenant that will own this agent.
                  examples: ["664f1b2c3e4d5a6b7c8d9e0f"]
                name:
                  type: string
                  description: Human-readable label for the agent.
                  examples: ["london-office-agent"]
                environment:
                  type: string
                  enum: [production, staging, development]
                  default: production
            examples:
              production:
                summary: Production agent
                value:
                  tenantId: "664f1b2c3e4d5a6b7c8d9e0f"
                  name: london-office-agent
                  environment: production
      responses:
        "201":
          description: Agent provisioned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: "#/components/schemas/Agent"
                  deploymentConfig:
                    $ref: "#/components/schemas/DeploymentConfig"
        "400":
          description: Missing required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Tenant not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Events ─────────────────────────────────────────────────────────────────
...